Complete rewrite of Rescue.
Finally does what I've wanted it to do. Managed to fit the entire
language with all options into the pattern. Can match forks in
heirarchies using the patterns. Ors are now done with an object acting
as a map and the patched result can be used in a switch statement.
Currying exists. Better error reporting for pattern errors is on the
way.